refactor(editablelist): move styling to its own file
